Refactor: 커스텀 훅 제거
