--- src/hooks/useResignateMember.ts
@@ -1,16 +1,0 @@
-export const useResignateMember = (
-  useResignate: any,
-  currentPostId: number,
-  setIsOpenedModal: React.Dispatch<React.SetStateAction<boolean>>,
-  nickname: string,
-) => {
-  const { mutate, isSuccess } = useResignate(currentPostId, nickname);
-
-  if (isSuccess) {
-    setIsOpenedModal(false);
-  }
-
-  return {
-    onClickResignateButton: mutate,
-  };
-};
